requirements.txt
streamlit
ultralytics
opencv-python-headless
ffmpeg-python
